refactor: Product/CartItem 가격 타입 BigDecimal 통일 (#50)
* refactor(product): Product 가격 타입 Long에서 BigDecimal로 변경
- Product 엔티티 price 필드 BigDecimal 전환 및 precision/scale 설정
- ProductCreateRequest, ProductUpdateRequest, ProductResponse DTO 반영
- ProductFixture 및 Product 관련 테스트 코드 수정

* refactor(cartItem): CartItem 가격 타입 Long에서 BigDecimal로 변경
- CartItem 엔티티 productPrice, totalPrice 필드 BigDecimal 전환
- totalPrice 계산 로직 BigDecimal.multiply 방식으로 변경
- CartItemResponse DTO 반영
- CartItemFixture 및 CartItem 관련 테스트 코드 수정

* refactor(order): OrderCommandServiceImpl BigDecimal 변환 코드 제거
- CartItem.productPrice가 BigDecimal이므로 불필요한 BigDecimal() 래핑 제거
- totalAmount 계산을 fold 방식으로 변경

+++ src/main/kotlin/com/hoppingmall/mall/cartItem/dto/response/CartItemResponse.kt
@@ -1,15 +1,16 @@
 package com.hoppingmall.mall.cartItem.dto.response
 
 import com.hoppingmall.mall.cartItem.domain.CartItem
+import java.math.BigDecimal
 
 data class CartItemResponse(
     val id: Long,
     val productId: Long,
     val productName: String,
-    val productPrice: Long,
+    val productPrice: BigDecimal,
     val productImageUrl: String?,
     val quantity: Int,
-    val totalPrice: Long
+    val totalPrice: BigDecimal
 ) {
     companion object {
         fun from(cartItem: CartItem): CartItemResponse {
